DOC: moved potentials out of tests and added derivatives doc
