+++ docs/matscipy.calculators.manybody.rst
@@ -1,0 +1,22 @@
+matscipy.calculators.manybody package
+=====================================
+
+.. automodule:: matscipy.calculators.manybody
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+Modules
+-------
+
+.. autosummary::
+    matscipy.calculators.manybody.explicit_forms
+
+matscipy.calculators.manybody.explicit_forms
+--------------------------------------------
+
+.. automodule:: matscipy.calculators.manybody.explicit_forms.harmonic
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :special-members:
